w: use perl from path
Note: native perl in Mojave is 5.18 which is not compatible w/ <<>>

The current version of Perl 5 is around 5.30.1.
No one should be running Perl 5.18. If you are on macOS,
use homebrew to get a better (safer) version of Perl.

---
You'd think that picking a perl from the path would be easy.
It isn't. `#!/usr/bin/env perl -w` doesn't behave as expected on
some systems, instead it tries to find a program named `perl -w`...

So, I'm using:
  https://perlbrew.pl/Dealing-with-shebangs.html